+++ examples/deliver_as.rs
@@ -3,6 +3,7 @@ use fama::PipeContent;
 
 #[tokio::main]
 async fn main() {
+    type AnswerResult = Result<String, String>;
     let counter = 0;
     let answer = fama::Pipeline::pass(counter)
         .through_fn(|mut c: i32, pipe: PipeContent| async move {
@@ -20,10 +21,12 @@ async fn main() {
         .through_fn(|c: i32, pipe: PipeContent| async move {
             // Change counter to a string
             pipe.container().set_type(c.to_string());
+            pipe.container()
+                .set_type(Ok::<String, String>(c.to_string()));
             Some(pipe)
         })
         .await
-        .deliver_as::<String>(); // request that a string be returned
+        .deliver_as::<AnswerResult>(); // request that a string be returned
 
     println!("example one answer is: {:?}", answer);