+++ src/pipeline.rs
@@ -2,11 +2,10 @@ use async_trait::async_trait;
 use busybody::ServiceContainer;
 use std::{future::Future, marker::PhantomData, sync::Arc};
 
-use crate::PipeContent;
+use crate::{content::PipeState, PipeContent};
 
 /// The pipes manager
 pub struct Pipeline<T: Send + Sync + 'static> {
-    fluid: PipeContent,
     phantom: PhantomData<T>,
     container: Arc<ServiceContainer>,
     went_through: bool,
@@ -16,17 +15,10 @@ impl<T: Clone + Send + Sync + 'static> Pipeline<T> {
     /// Accepts the pipeline content/input.
     /// This is the beginning of the pipeline
     pub fn pass(content: T) -> Self {
-        let mut fluid = PipeContent::default();
-
-        let container = Arc::new(ServiceContainer::proxy());
-        container.set_type(content);
-
-        fluid.1 = Some(container.clone());
-        container.set_type(fluid.clone());
+        let fluid = PipeContent::new(content);
 
         Self {
-            fluid,
-            container,
+            container: fluid.container().clone(),
             phantom: PhantomData,
             went_through: false,
         }
@@ -36,16 +28,54 @@ impl<T: Clone + Send + Sync + 'static> Pipeline<T> {
     /// The closure can accept zero or more arguments.
     /// Unlike a struct pipe, a closure does not have to use a tuple
     /// for multiple arguments. Arguments can be up to 17
-    pub async fn through_fn<P, Args>(mut self, pipe: P) -> Self
+    pub async fn through_fn<H, Args, O>(mut self, handler: H) -> Self
     where
-        P: PipeFn<Args>,
+        H: PipeFnHandler<Args, O>,
         Args: busybody::Injectable + 'static,
     {
-        if !self.fluid.0 {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
             let args = Args::inject(&self.container).await;
-            if let Some(f) = pipe.call(args).await {
-                self.fluid = f;
+            handler.call(args).await;
+            self.went_through = true;
+        } else {
+            self.went_through = false;
+        }
+
+        self
+    }
+
+    /// Accepts a closure or function as a pipe.
+    /// The closure can accept zero or more arguments.
+    /// Unlike a struct pipe, a closure does not have to use a tuple
+    /// for multiple arguments. Arguments can be up to 17
+    /// Closure must return a boolean. `False` will stop the pipe flow
+    pub async fn next_fn<H, Args>(mut self, handler: H) -> Self
+    where
+        H: PipeFnHandler<Args, bool>,
+        Args: busybody::Injectable + 'static,
+    {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
+            let args = Args::inject(&self.container).await;
+            if !handler.call(args).await {
+                self.container.set(PipeState::Stop);
             }
+            self.went_through = true;
+        } else {
+            self.went_through = false;
+        }
+
+        self
+    }
+
+    /// Stores the result from the pipe handler
+    pub async fn store_fn<H, Args, O: Clone + Send + Sync + 'static>(mut self, handler: H) -> Self
+    where
+        H: PipeFnHandler<Args, O>,
+        Args: busybody::Injectable + 'static,
+    {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
+            let args = Args::inject(&self.container).await;
+            self.container.set_type(handler.call(args).await);
             self.went_through = true;
         } else {
             self.went_through = false;
@@ -55,16 +85,51 @@ impl<T: Clone + Send + Sync + 'static> Pipeline<T> {
     }
 
     /// Accepts an instance of a struct that implements `fama::FamaPipe`
-    pub async fn through<P, Args>(mut self, pipe: P) -> Self
+    /// The returned result will be store for the next pipe handlers
+    pub async fn through<H, Args, O>(mut self, handler: H) -> Self
     where
-        P: FamaPipe<Args>,
+        H: FamaPipe<Args, O>,
         Args: busybody::Injectable + 'static,
     {
-        if !self.fluid.0 {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
             let args = Args::inject(&self.container).await;
-            if let Some(f) = pipe.receive_pipe_content(args).await {
-                self.fluid = f
+            handler.receive_pipe_content(args).await;
+            self.went_through = true;
+        } else {
+            self.went_through = false;
+        }
+
+        self
+    }
+
+    /// Accepts an instance of a struct that implements `fama::FamaPipe`
+    /// Must return a boolean. `False` will halt the flow
+    pub async fn next<H, Args>(mut self, handler: H) -> Self
+    where
+        H: FamaPipe<Args, bool>,
+        Args: busybody::Injectable + 'static,
+    {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
+            let args = Args::inject(&self.container).await;
+            if !handler.receive_pipe_content(args).await {
+                self.container.set(PipeState::Stop);
             }
+            self.went_through = true;
+        } else {
+            self.went_through = false;
+        }
+
+        self
+    }
+    pub async fn store<H, Args, O: Clone + Send + Sync + 'static>(mut self, handler: H) -> Self
+    where
+        H: FamaPipe<Args, O>,
+        Args: busybody::Injectable + 'static,
+    {
+        if *self.container.get::<PipeState>().unwrap() == PipeState::Run {
+            let args = Args::inject(&self.container).await;
+            self.container
+                .set_type(handler.receive_pipe_content(args).await);
             self.went_through = true;
         } else {
             self.went_through = false;
@@ -74,14 +139,22 @@ impl<T: Clone + Send + Sync + 'static> Pipeline<T> {
     }
 
     /// Returns the passed variable
-    pub fn deliver(self) -> T {
+    pub fn deliver(&self) -> T {
         self.container.get_type().unwrap()
+    }
+
+    pub fn try_to_deliver(&self) -> Option<T> {
+        self.container.get_type()
     }
 
     /// Returns a different type that may have been set
     /// by one of the pipes
-    pub fn deliver_as<R: Clone + 'static>(self) -> R {
+    pub fn deliver_as<R: Clone + 'static>(&self) -> R {
         self.container.get_type().unwrap()
+    }
+
+    pub fn try_deliver_as<R: Clone + 'static>(&self) -> Option<R> {
+        self.container.get_type()
     }
 
     /// Returns true if the content went through all the registered pipes
@@ -91,9 +164,9 @@ impl<T: Clone + Send + Sync + 'static> Pipeline<T> {
 }
 
 #[async_trait]
-pub trait FamaPipe<Args = ()> {
+pub trait FamaPipe<Args, O> {
     /// Where a pipe logic resides
-    async fn receive_pipe_content(&self, args: Args) -> Option<PipeContent>;
+    async fn receive_pipe_content(&self, args: Args) -> O;
 
     /// Wraps the type in a Box
     fn to_pipe(self) -> Box<Self>
@@ -104,16 +177,16 @@ pub trait FamaPipe<Args = ()> {
     }
 }
 
-pub trait PipeFn<Args>: Send + Sync + 'static {
-    type Future: Future<Output = Option<PipeContent>> + Send;
+pub trait PipeFnHandler<Args, O>: Send + Sync + 'static {
+    type Future: Future<Output = O> + Send;
 
     fn call(&self, args: Args) -> Self::Future;
 }
 
-impl<Func, Fut> PipeFn<()> for Func
+impl<Func, Fut, O> PipeFnHandler<(), O> for Func
 where
     Func: Send + Sync + Fn() -> Fut + Send + Sync + 'static,
-    Fut: Future<Output = Option<PipeContent>> + Send,
+    Fut: Future<Output = O> + Send,
 {
     type Future = Fut;
     fn call(&self, _: ()) -> Self::Future {
@@ -121,10 +194,10 @@ where
     }
 }
 
-impl<Func, Arg1, Fut> PipeFn<(Arg1,)> for Func
+impl<Func, Arg1, Fut, O> PipeFnHandler<(Arg1,), O> for Func
 where
     Func: Send + Sync + Fn(Arg1) -> Fut + Send + Sync + 'static,
-    Fut: Future<Output = Option<PipeContent>> + Send,
+    Fut: Future<Output = O> + Send,
 {
     type Future = Fut;
     fn call(&self, (c,): (Arg1,)) -> Self::Future {
@@ -134,9 +207,9 @@ where
 
 macro_rules! pipe_func{
     ($($T: ident),*) => {
-        impl<Func, $($T),+, Fut> PipeFn <($($T),+)> for Func
+        impl<Func, $($T),+, Fut, O> PipeFnHandler <($($T),+), O> for Func
          where Func: Fn($($T),+) -> Fut + Send + Sync + 'static,
-         Fut: Future<Output = Option<PipeContent>> + Send,
+         Fut: Future<Output = O> + Send,
         {
             type Future = Fut;
 
@@ -164,3 +237,180 @@ pipe_func! {Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10, Arg11, 
 pipe_func! {Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10, Arg11, Arg12, Arg13, Arg14, Arg15}
 pipe_func! {Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10, Arg11, Arg12, Arg13, Arg14, Arg15, Arg16}
 pipe_func! {Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10, Arg11, Arg12, Arg13, Arg14, Arg15, Arg16, Arg17}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    struct AddOne;
+    #[async_trait]
+    impl FamaPipe<i32, i32> for AddOne {
+        async fn receive_pipe_content(&self, num: i32) -> i32 {
+            num + 1
+        }
+    }
+
+    struct AddTwo;
+    #[async_trait]
+    impl FamaPipe<i32, i32> for AddTwo {
+        async fn receive_pipe_content(&self, num: i32) -> i32 {
+            num + 2
+        }
+    }
+
+    struct StoreAddOne;
+    #[async_trait]
+    impl FamaPipe<(i32, PipeContent), ()> for StoreAddOne {
+        async fn receive_pipe_content(&self, (num, pipe): (i32, PipeContent)) {
+            pipe.store(num + 1);
+        }
+    }
+
+    struct StoreAddTwo;
+    #[async_trait]
+    impl FamaPipe<(i32, PipeContent), ()> for StoreAddTwo {
+        async fn receive_pipe_content(&self, (num, pipe): (i32, PipeContent)) {
+            pipe.store(num + 2);
+        }
+    }
+
+    struct ValidateCount;
+    #[async_trait]
+    impl FamaPipe<i32, bool> for ValidateCount {
+        async fn receive_pipe_content(&self, num: i32) -> bool {
+            num >= 6
+        }
+    }
+
+    #[tokio::test]
+    async fn test_through() {
+        let result = Pipeline::pass(0)
+            .through(StoreAddOne)
+            .await
+            .through(StoreAddOne)
+            .await
+            .through(StoreAddTwo)
+            .await
+            .through(StoreAddTwo)
+            .await
+            .deliver();
+
+        assert_eq!(result, 6);
+    }
+
+    #[tokio::test]
+    async fn test_store() {
+        let result = Pipeline::pass(0)
+            .store(AddOne)
+            .await
+            .store(AddOne)
+            .await
+            .store(AddTwo)
+            .await
+            .store(AddTwo)
+            .await
+            .deliver();
+
+        assert_eq!(result, 6);
+    }
+
+    #[tokio::test]
+    async fn test_next() {
+        let result = Pipeline::pass(0)
+            .store(AddOne)
+            .await
+            .store(AddOne)
+            .await
+            .store(AddTwo)
+            .await
+            .next(ValidateCount)
+            .await
+            .store(AddTwo)
+            .await
+            .deliver();
+
+        assert_eq!(result, 4);
+
+        let result = Pipeline::pass(0)
+            .store(AddOne)
+            .await
+            .store(AddOne)
+            .await
+            .store(AddTwo)
+            .await
+            .store(AddTwo)
+            .await
+            .store(AddTwo)
+            .await
+            .next(ValidateCount)
+            .await
+            .store(AddTwo)
+            .await
+            .deliver();
+
+        assert_eq!(result, 10);
+    }
+
+    #[tokio::test]
+    async fn test_through_fn1() {
+        let result: bool = Pipeline::pass(33)
+            .through_fn(|num: i32, pipe: PipeContent| async move {
+                pipe.store(num + 2);
+            })
+            .await
+            .through_fn(|num: i32, pipe: PipeContent| async move {
+                pipe.store(num == 35);
+            })
+            .await
+            .deliver_as();
+
+        assert_eq!(result, true);
+    }
+
+    #[tokio::test]
+    async fn test_next_fn() {
+        let result: bool = Pipeline::pass(33)
+            .next_fn(|num: i32, pipe: PipeContent| async move {
+                pipe.store(num + 2);
+                true
+            })
+            .await
+            .next_fn(|num: i32| async move { num != 35 })
+            .await
+            .next_fn(|num: i32| async move { num == 35 })
+            .await
+            .confirm();
+
+        assert_eq!(result, false);
+    }
+
+    #[tokio::test]
+    async fn test_store_fn() {
+        let total = Pipeline::pass(0)
+            .store_fn(|num: i32| async move { num + 1 })
+            .await
+            .store_fn(|num: i32| async move { num + 4 })
+            .await
+            .store_fn(|num: i32| async move { num * 5 })
+            .await
+            .deliver();
+
+        assert_eq!(total, 25);
+    }
+
+    #[tokio::test]
+    async fn test_deliver_as() {
+        let result: bool = Pipeline::pass(0)
+            .store_fn(|num: i32| async move { num + 1 })
+            .await
+            .store_fn(|num: i32| async move { num + 4 })
+            .await
+            .store_fn(|num: i32| async move { num * 5 })
+            .await
+            .store_fn(|num: i32| async move { num == 25 })
+            .await
+            .deliver_as();
+
+        assert_eq!(result, true);
+    }
+}