+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change logs
 
+## Version 0.3.0
+Pipe handlers can now specify returned value. This allows you to use general function or struct instead
+of specific ones. 
+
 ## Version 0.2.0
-The entire crate was reimplemented. `Busyboby` is use for dependency injecting.
+The entire crate was reimplemented. `Busybody` is use for dependency injecting.